broken version

+++ lib/db.py
@@ -1,15 +1,17 @@
 import mysql.connector as mysql
+import logging
 
+logging.getLogger(__name__)
 
 class DB:
     """
     Base class that implements database connections
     """
-    def __new__(cls, *args, **kwargs):
-        """ Implement as singleton"""
-        if not hasattr(cls, 'instance'):
-            cls.instance = super().__new__(cls)
-        return cls.instance
+    # def __new__(cls, *args, **kwargs):
+    #     """ Implement as singleton"""
+    #     if not hasattr(cls, 'instance'):
+    #         cls.instance = super().__new__(cls)
+    #     return cls.instance
 
     def __init__(self,
                  database: str = None,
@@ -22,18 +24,19 @@ class DB:
             auth_plugin = 'mysql_native_password',
             autocommit = True
         )
-
         self.curs = self.conn.cursor(buffered = True, dictionary = True)
 
     def query(self, query: str, results: bool = True):
         """ Execute a query into the selected database."""
         self.curs.execute(query)
+        logging.debug(f"SQL: ``{query}`")
         if results:
             return self.curs.fetchall()
 
     def change_db(self, database: str):
         """ Select the database. Equivalent of `use <DATABASE>`"""
         self.conn.database = database
+        logging.debug(f"Changing DB to {database}")
 
     def select_where(self, table, *columns, **wheres) -> dict or list:
         """
@@ -57,6 +60,7 @@ class DB:
             sql = sql[:-2] # trim final comma
         print(sql)
         ret = self.query(sql)
+        logging.debug(f"SQL: {sql}")
         if multi:
             return ret
         else:
@@ -69,6 +73,7 @@ class DB:
         """
         sql = f"SELECT {column} from {table}"
 
+        logging.debug(f"SQL: {sql}")
         if multi:
             return self.query(sql)
         else:
@@ -80,6 +85,7 @@ class DB:
         else returns a dictionary.
         """
         sql = f"SELECT * FROM {table} WHERE {key} = '{value}'"
+        logging.debug(f"SQL: {sql}")
         return self.query(sql)
 
     def insert_row(self, table, **kwargs):
@@ -95,8 +101,8 @@ class DB:
             sql += f"'{v}', "
         sql = sql[:-2] # trim ", "
         sql += ")"
-        print(sql)
         self.query(sql, results = False)
+        logging.debug(f"SQL: {sql}")
         return True
 
     def update_value(self, table, key, new_value, **wheres):
@@ -112,6 +118,7 @@ class DB:
             sql = sql[:-6] # cleave final 'and '
         print(sql)
         self.query(sql, results = False)
+        logging.debug(f"SQL: {sql}")
         return True
 
     def update_row(self, table, selector, value, **new_values):
@@ -122,7 +129,7 @@ class DB:
         sql += f" WHERE {selector} = '{value}'"
 
         self.query(sql, results = False)
-
+        logging.debug(f"SQL: {sql}")
         return True
 
     def get_table_names(self) -> list:
@@ -130,6 +137,7 @@ class DB:
         Get the names of each table, as a list.
         """
         ret = self.query("SHOW TABLES")
+        logging.debug("SQL: SHOW TABLES")
         names = []
         for item in ret:
             names.append(list(item.values())[0])
@@ -140,15 +148,25 @@ class DB:
         Delete a row from table, based on wheres
         """
 
+        if not wheres:
+            raise Exception("Missing `wheres` kwargs in call to delete_row")
         sql = f"DELETE FROM {table} WHERE "
+        for k, v in wheres.items():
+            sql += f"{k} = {v}, "
+        sql = sql[:2]
+        logging.debug(f"SQL: {sql}")
+
 
     def select_all(self, table_name):
-        ret = self.query(f"SELECT * FROM {table_name}")
+        sql = f"SELECT * FROM {table_name}"
+        ret = self.query(sql)
+        logging.debug(f"SQL: {sql}")
         return ret
 
     def get_columns_names(self, table) -> list:
         sql = f"describe {table}"
         ret = self.query(sql)
+        logging.debug(f"SQL: {sql}")
         names = []
         for field in ret:
             names.append(field.get("Field"))
